pyproject.toml
[tool.poetry]
name = "api-central"
version = "0.1.0"
description = "Bridges Market Central API"
authors = ["marxxt <marxxt@me.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {extras = ["fastapi"], version = "^0.270.2"}
uvicorn = "^0.34.2"
jwt = "^1.3.1"
supabase = "^2.15.2"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.1.0"
google-cloud-firestore = "2.20.2"
redis = "^6.1.0"
email-validator = "^2.2.0"
motor = "^3.7.1"
faker = "^37.3.0"
celery = "^5.4.0" # Add Celery dependency
aioboto3 = "^12.0.0" # For SQS if needed, but generally good for async AWS
phonenumbers = "^9.0.5"
pydantic-extra-types = "^2.10.4"
faker-crypto = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
pytest-mock = "^3.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
